ci: Add initial scale test workflow for bootstrapping
This commit acts as an initial effort to add automated scale testing to
Cilium, as described in https://github.com/cilium/design-cfps/pull/11.
The workflow added as part of this commit is meant to lay down a
foundation that can be built upon, with the intent that it will be
iteratively developed on to carry out the CFP. The workflow currently
performs the following:

* Create a two-node GCE cluster using kops
* Run a bare-bones ClusterLoader2 configuration
* Cleanup the cluster

The following next steps are a WIP:

* Expand on the ClusterLoader2 configuration
* Increase the number of nodes in the cluster
* Export results from ClusterLoader2
* Allow for manual runs using WorkflowDispatch

Note: ClusterLoader2 is currently pulled from a fork containing changes
that are intended to be merged upstream in the coming week.

Fixes: #27749

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>